+++ Plugin.cs
@@ -16,7 +16,7 @@ namespace OttoStash;
 public class OttoStashPlugin : BaseUnityPlugin
 {
     internal const string ModName = "OttoStash";
-    internal const string ModVersion = "3.1.1";
+    internal const string ModVersion = "3.1.2";
     internal const string Author = "potto007";
     internal const string ModGUID = $"{Author}.{ModName}";
     internal const string KgGuid = "kg.ItemDrawers";